+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
   * gpxdo clear --untangle
   * gpxdo help, gpxdo help ls
   * new backend Memory
+  * gpxdo with colors
 
 1.6.0 release 2019-01-31
 ------------------------